requirements.txt
torch
open-clip-torch
faiss-cpu
pillow
numpy
easyocr
ultralytics
